# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/search": {
      "post": {
        "description": "Results are cached per normalized query (see SEARCH_CACHE_TTL_SECONDS). With `summarize: true` an AI provider writes a grounded answer with [n] citations; this consumes one AI request from the plan quota. If only the summary fails, results are still returned with `summaryError` set.",
        "operationId": "Search_perform",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQueryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformSearchResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource conflicts with existing state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit or plan quota exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "502": {
            "description": "Upstream AI / search provider failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Search the web (optionally with an AI summary)",
        "tags": [
          "Search"
        ]
      }
    },
    "/api/v1/search/history": {
      "get": {
        "operationId": "Search_history",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Page size (max 100).",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Filter history by query text (case-insensitive).",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SearchResponseDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Search history (newest first)",
        "tags": [
          "Search"
        ]
      },
      "delete": {
        "operationId": "Search_clearHistory",
        "parameters": [],
        "responses": {
          "204": {
            "description": "History cleared"
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Clear the entire search history",
        "tags": [
          "Search"
        ]
      }
    },
    "/api/v1/search/history/{id}": {
      "get": {
        "operationId": "Search_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a past search with its results",
        "tags": [
          "Search"
        ]
      },
      "delete": {
        "operationId": "Search_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Search deleted"
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete one search from history",
        "tags": [
          "Search"
        ]
      }
    },
    "/api/v1/search/recent": {
      "get": {
        "operationId": "Search_recent",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Number of distinct recent queries (max 50).",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecentSearchDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Recent distinct searches of the current user",
        "tags": [
          "Search"
        ]
      }
    },
    "/api/v1/search/suggestions": {
      "get": {
        "description": "Combines the user’s own matching history (most frequent first) with engine autocomplete. Other users’ queries are never suggested.",
        "operationId": "Search_suggestions",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Prefix typed so far.",
            "schema": {
              "minLength": 1,
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Max suggestions (max 20).",
            "schema": {
              "minimum": 1,
              "maximum": 20,
              "default": 8,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Search suggestions for a prefix",
        "tags": [
          "Search"
        ]
      }
    },
...
  },
  "info": {
    "title": "EchoGPT API",
    "description": "\nREST API powering the **EchoGPT** multi-AI chat Chrome extension.\n\n### Authentication\n1. `POST /api/v1/auth/register` or `POST /api/v1/auth/login` returns an **access token** (short lived JWT) and a **refresh token** (opaque, long lived, rotated on every use).\n2. Send the access token as `Authorization: Bearer <token>`. Click **Authorize** above to use it here.\n3. When the access token expires call `POST /api/v1/auth/refresh`. Re-using an already rotated refresh token revokes the whole session.\n\n### Errors\nEvery error uses the same envelope: `{ statusCode, error, message, details?, path, timestamp, requestId }`.\nThe `x-request-id` response header echoes the correlation id for support/debugging.\n\n### Quotas\nAI endpoints (chat, AI-assisted search) consume the caller's daily plan quota. Exceeding it returns **429**.\n",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "details": {
            "example": [
              "email must be an email",
              "password must be longer than or equal to 8 characters"
            ],
            "description": "Field level details, present for validation errors.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "path": {
            "type": "string",
            "example": "/api/v1/auth/register"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-29T10:15:30.000Z"
          },
          "requestId": {
            "type": "string",
            "example": "6f1c2c1e-6c47-4f0e-9a36-0c5f0f4b1a2d"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp",
          "requestId"
        ]
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 57
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
...
      "SearchQueryDto": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 500,
            "example": "What is NestJS?"
          },
          "limit": {
            "type": "number",
            "default": 5,
            "minimum": 1,
            "maximum": 10
          },
          "summarize": {
            "type": "boolean",
            "default": false,
            "description": "Generate an AI answer grounded in the results (with [n] citations). Consumes one AI request from the plan quota."
          },
          "providerId": {
            "type": "string",
            "format": "uuid",
            "description": "Provider for the summary (default if omitted)"
          },
          "model": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "gpt-5-mini",
            "description": "Model for the summary"
          }
        },
        "required": [
          "query"
        ]
      },
      "SearchResultItemDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "NestJS"
          },
          "url": {
            "type": "string",
            "example": "https://en.wikipedia.org/wiki/NestJS"
          },
          "snippet": {
            "type": "string",
            "example": "NestJS is a server-side Node.js web framework..."
          },
          "source": {
            "type": "string",
            "example": "Wikipedia"
          }
        },
        "required": [
          "title",
          "url",
          "snippet"
        ]
      },
      "PerformSearchResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "History entry id"
          },
          "query": {
            "type": "string",
            "example": "What is NestJS?"
          },
          "engine": {
            "type": "string",
            "example": "duckduckgo"
          },
          "resultCount": {
            "type": "number",
            "example": 5
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResultItemDto"
            }
          },
          "fromCache": {
            "type": "boolean",
            "description": "Results were served from the search cache"
          },
          "aiSummary": {
            "type": "string",
            "nullable": true,
            "example": "NestJS is a Node.js framework for building server-side applications [1]."
          },
          "providerId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "summaryError": {
            "type": "string",
            "nullable": true,
            "description": "Why the AI summary is missing although it was requested (results still returned).",
            "example": null
          }
        },
        "required": [
          "id",
          "query",
          "engine",
          "resultCount",
          "results",
          "fromCache",
          "aiSummary",
          "providerId",
          "createdAt",
          "summaryError"
        ]
      },
      "SearchResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "History entry id"
          },
          "query": {
            "type": "string",
            "example": "What is NestJS?"
          },
          "engine": {
            "type": "string",
            "example": "duckduckgo"
          },
          "resultCount": {
            "type": "number",
            "example": 5
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResultItemDto"
            }
          },
          "fromCache": {
            "type": "boolean",
            "description": "Results were served from the search cache"
          },
          "aiSummary": {
            "type": "string",
            "nullable": true,
            "example": "NestJS is a Node.js framework for building server-side applications [1]."
          },
          "providerId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "query",
          "engine",
          "resultCount",
          "results",
          "fromCache",
          "aiSummary",
          "providerId",
          "createdAt"
        ]
      },
      "RecentSearchDto": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "example": "what is nestjs?"
          },
          "count": {
            "type": "number",
            "example": 3
          },
          "lastSearchedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "query",
          "count",
          "lastSearchedAt"
        ]
      },
      "SuggestionsResponseDto": {
        "type": "object",
        "properties": {
          "suggestions": {
            "example": [
              "nestjs",
              "nestjs prisma",
              "nestjs tutorial"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "suggestions"
        ]
      },
...
    }
  }
}